Replace model wrapper with a configurable model_transform.
