+++ mart/callbacks/adversarial_training.py
@@ -11,8 +11,6 @@ from typing import Callable
 
 from lightning.pytorch.callbacks import Callback
 
-from ..utils import MonkeyPatch
-
 __all__ = ["AdversarialTraining"]
 
 
@@ -48,28 +46,6 @@ class AdversarialTraining(Callback):
     def teardown(self, trainer, pl_module, stage=None):
         pl_module.on_after_batch_transfer = self._on_after_batch_transfer
 
-    def wrap_model(self, model, dataloader_idx):
-        """Make a model, such that `output = model(batch)`."""
-
-        # Consume dataloader_idx
-        if hasattr(model, "attack_step"):
-
-            def model_forward(batch):
-                output = model.attack_step(batch, dataloader_idx)
-                return output
-
-        elif hasattr(model, "training_step"):
-            # Monkey-patch model.log to avoid spamming.
-            def model_forward(batch):
-                with MonkeyPatch(model, "log", lambda *args, **kwargs: None):
-                    output = model.training_step(batch, dataloader_idx)
-                return output
-
-        else:
-            model_forward = model
-
-        return model_forward
-
     def on_after_batch_transfer(self, pl_module, batch, dataloader_idx):
         batch = self._on_after_batch_transfer(batch, dataloader_idx)
 
@@ -90,12 +66,7 @@ class AdversarialTraining(Callback):
         # Move adversary to same device as pl_module and run attack
         adversary.to(pl_module.device)
 
-        # We assume Adversary is not aware of PyTorch Lightning,
-        # so wrap the model as `output=model(batch)`.
-        model = self.wrap_model(pl_module, dataloader_idx)
-
-        # TODO: We may need to do model.eval() if there's BN-like layers in the model.
         # Directly pass batch instead of assuming it has a structure.
-        batch_adv = adversary(batch=batch, model=model)
+        batch_adv = adversary(batch=batch, model=pl_module)
 
         return batch_adv
