+++ mart/configs/attack/adversary.yaml
@@ -1,5 +1,6 @@
 defaults:
   - /callbacks@callbacks: [progress_bar]
+  - model_transform: lightning
 
 _target_: mart.attack.Adversary
 _convert_: all
