+++ mart/configs/attack/model_transform/lightning.yaml
@@ -1,0 +1,1 @@
+_target_: mart.attack.model_transform.LightningModuleAsTarget
